+++ yospace/src/main/java/com/bitmovin/player/integration/yospace/BitmovinYospacePlayer.kt
@@ -20,10 +20,33 @@ import com.bitmovin.player.api.metadata.emsg.EventMessage
 import com.bitmovin.player.api.metadata.id3.BinaryFrame
 import com.bitmovin.player.api.source.SourceType as MediaSourceType
 import com.bitmovin.player.api.source.*
+import com.bitmovin.player.integration.yospace.advertising.Ad
+import com.bitmovin.player.integration.yospace.advertising.AdBreak
+import com.bitmovin.player.integration.yospace.advertising.AdBreakPosition
+import com.bitmovin.player.integration.yospace.advertising.AdTimeline
+import com.bitmovin.player.integration.yospace.advertising.BitmovinTruexAdRenderer
+import com.bitmovin.player.integration.yospace.advertising.BitmovinTruexAdRendererListener
+import com.bitmovin.player.integration.yospace.advertising.CompanionAd
+import com.bitmovin.player.integration.yospace.advertising.CompanionAdResource
+import com.bitmovin.player.integration.yospace.advertising.CompanionAdType
 import com.bitmovin.player.integration.yospace.config.TruexConfig
+import com.bitmovin.player.integration.yospace.config.YospaceAssetType
 import com.bitmovin.player.integration.yospace.config.YospaceConfig
 import com.bitmovin.player.integration.yospace.config.YospaceDebugMode
+import com.bitmovin.player.integration.yospace.config.YospaceLiveInitializationType
 import com.bitmovin.player.integration.yospace.config.YospaceSourceConfig
+import com.bitmovin.player.integration.yospace.deficiency.YospaceErrorCode
+import com.bitmovin.player.integration.yospace.deficiency.YospaceWarningCode
+import com.bitmovin.player.integration.yospace.policy.BitmovinYospacePlayerPolicy
+import com.bitmovin.player.integration.yospace.policy.DefaultBitmovinYospacePlayerPolicy
+import com.bitmovin.player.integration.yospace.policy.YospacePlayerPolicy
+import com.bitmovin.player.integration.yospace.events.AdClickThroughReporter
+import com.bitmovin.player.integration.yospace.events.AdjustedTimeChangedDispatcher
+import com.bitmovin.player.integration.yospace.events.YospaceAdStartedSnapshot
+import com.bitmovin.player.integration.yospace.events.YospaceEventEmitter
+import com.bitmovin.player.integration.yospace.events.YospacePlayerEvent
+import com.bitmovin.player.integration.yospace.events.YospacePlayerEventDispatcher
+import com.bitmovin.player.integration.yospace.events.YospacePlayerEventListener
 import com.yospace.admanagement.*
 import com.yospace.admanagement.TimedMetadata
 import com.yospace.admanagement.EventListener as YospaceEventListener