+++ yospace/src/main/java/com/bitmovin/player/integration/yospace/advertising/AdData.kt
@@ -1,4 +1,4 @@
-package com.bitmovin.player.integration.yospace
+package com.bitmovin.player.integration.yospace.advertising
 
 data class AdData(
     override val mimeType: String,