+++ yospace/src/main/java/com/bitmovin/player/integration/yospace/config/YospaceSourceConfig.kt
@@ -1,7 +1,5 @@
 package com.bitmovin.player.integration.yospace.config
 
-import com.bitmovin.player.integration.yospace.YospaceAssetType
-
 data class YospaceSourceConfig(
     val assetType: YospaceAssetType,
     val retryExcludingYospace: Boolean = false