+++ yospace/src/main/java/com/bitmovin/player/integration/yospace/events/YospacePlayerEventDispatcher.kt
@@ -1,4 +1,4 @@
-package com.bitmovin.player.integration.yospace
+package com.bitmovin.player.integration.yospace.events
 
 import kotlin.reflect.KClass